hw4/user_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==22.0.0
sqlalchemy==2.0.35
pymysql==1.1.1
cryptography==43.0.1
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
pydantic[email]==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
authlib==1.3.2

# Azure
azure-identity==1.18.0
azure-keyvault-secrets==4.8.0
azure-servicebus==7.12.3
azure-cosmos==4.7.0
